Added openCVManager code to JewelTest
Now throws a NullPointerException on  the parameter for the BaseLoaderCallback. Need to find how to access the app context for the JewelTest class. this.hardwaremap.appContext, which is the parameter you are supposed to give to access the file's app context, but for some reason, it returns null... so ya. :(