Changes to Jewel Test
Addition comments, and created an if-else statement to prevent the code from crashing if image fails to load. Found out that the error is caused by the code taking too long to respond, causing the opmode to freak out and shut down the loop(). Need to figure out how to stop this from occuring.