Made jewel detection a bit more accurate.
From ~1 ft away, the phone will pretty accurately detect the jewel as long as there are no distractions/other circles in the background (a white/silver background works quite well). There is one problem though. The phone will not detect the phone if there is not a bright light above it, and this only works if the light is ~0.5 ft above the jewel. Not sure how this could be worked around, the contrast between the jewel  and its surroundings needs to be extremely obvious.