Updates to Jewel Detect
Now, the program is able to detect the circles, identify there colors, AND tell us which side each circle is on. The detection part definitly needs to be fine-tuned, but besides that, it should be good to go. It would be wise to fine-tune the detection parameters after we finish the robot. That way we could test the parameters from the robot.